requirements.txt
biopython>=1.65
BeautifulSoup4>=4.3.2
argparse>=1.2.1
numpy>=1.9.1
airr>=1.3.0
